Creates new class for clearer implementation (#2)
* Implements MVP for Miniworld Sawmill Display

This application utilizes `omxplayer` and `pysimpledmx` to play video and control DMX lighting respectively.

The OMXPlayer wrapper communicates from the application using D-Bus to the OMXPlayer application which runs independently and is responsible for playing the video file.

The `pysimpledmx` library (which has been extended lightly in a limbic repo, hopefully soon to be merged upstream) communicates with an Enttec USB-to-DMX device. A light mock of this class has been added in case that device is not available.

The core of this application is to manage the video and dmx state based on a button press.
When the application starts, everything (video and dmx stuff) is setup and paused. When the button is pressed, the video and dmx sequence is started and will continue until the video is finished. The button does nothing while the video is running.

Other Classes:
ButtonHandler Class:
* Wraps GPIO library for async button with debounce functionality (i.e. prevents button from getting multiple hits in a short time).

Additionally:
* adds linux requirements script, will be the definitive source of RPI system state for this project
    * in the future this could be replaced with Ansible Playbook config.
* adds config.py for application configuration

* Creates new class for clearer implementation
removes large amounts of code from main

All of the functionality for running video/dmx has been moved to its own class (omxdmx, a subclass of Thread). This allows the main sawmill.py file to be much cleaner.

* added requirements for python

* changed name of callback function in buttonhandler as suggested by reviewer

* added f-string mod to omxdmx to try it out

* typo

* referenced a non-existing method

* removed unused variable in lambda function

* renamed ambiguous variable name

* better description for argparser

* renamed thread

* removed need to call buttonEvent.set() when killing thread

* separated exceptions into two cases

* exception in playerFactory static method now kill app