Fixes automatic stopping of mediafile if no SEQUENCE provided in config
When the `sequence` variable is not set when constructing OmxDmx (i.e. `sequence=[]`), a sequence was generated with an `end_time` equal to the length of the media file. This was causing the program to exit because the mediafile would end before the final value.

This commit adds a global variable to Omxdmx which offsets the automatically generated sequence `end_time` to be the length of the media file + this offset (a negative value).