Fixes issue with Omxplayer dying (finished playing) while Omxdmx thread asleep
A sequence can be created which does not finish before the media file. The original check for this was not effective at catching the issue. This commit implements a fix that checks if the `end_time` of the current sequence element will cause the video file to run out before the thread is finished sleeping. If this is the case, it logs the issue, pauses the media file, and skips the sequence.

**Note**: This is perhaps not the best strategy. It punishes the user for not making a correct sequence. Traversing the sequence and making sure nothing is longer than the media file could work better. Or, letting the file play until the last moment, so that it doesn't abruptly stop (which may feel like a bug).