Fixes a problem in the omxPlayerMock class
Previously, the omxPlayerMock class was setup to have the duration() value be +1 greater than position(). This no longer worked with the fact that the `steps['end_time']` was now used as the metric to find if the "media file" was finished.

This commit changes the returned value from `duration()` now returns a huge value: `sys.maxsize`.