Fixes omxPlayerMock position() to track elasped time to return a varying position value
Previously, the `position()` method would return a static position value. This did not interact properly with recent changes. This commit changes `position()` to keep track of elapsed time in order to properly emulate playing a media file and allow the application to run as if a media file was actually playing.