Adds toggle switch to control AUTOREPEAT
Previously, the AUTOREPEAT variable in the config file was the only way to control autorepeat. However, sometimes staying on the default dmx value is needed. A specific example is a "service" mode where all the lights turn on so work can be done on an install (instead of chaging the config file to have a static light show). This commit fixes this by:

* adding a GPIO swtich (and config) to modify the autorepeat state at boot (in app.py)
* changing where in the code the 'default' lighting level is set. This allows autorepeat to go on forever, never hitting the default lighting values BUT holding on the default when no activity.