Adds ability to have multiple values for default dmx
Previously the default dmx value was a single digit and all channels would be set to this value. However, since the addition of human input for setting AUTOREPEAT--which allows users to set the AUTOREPEAT state on startup with a toggle switch--it is valuable to be able to set a specific default configuration with more control over channel values. This commit adds the ability to have multiple values in the config to set multiple channels' default values.

A single int will still automatically map to all channels. A list with multiple values can now be used. The number of default values in the list can be less than the number of mapped channels. If the list is too short, it will be padded with zeros. If the list is too long, the extra values will be split off the end.

additionally:
* adds modification to example config to reflect new option for default values